Teach testcase in driver_suite the new prototype for driver_async
The testcase core dunmped. Also made sure the key could actually store
long integers in driver_async (which is more of a cosmetic change).

+++ erts/emulator/beam/erl_async.c
@@ -397,7 +397,7 @@ long driver_async(ErlDrvPort ix, unsigned long* key,
     ErlAsync* a = (ErlAsync*) erts_alloc(ERTS_ALC_T_ASYNC, sizeof(ErlAsync));
     Port* prt = erts_drvport2port(ix);
     long id;
-    unsigned int qix;
+    unsigned long qix;
 
 
     if (!prt)
